fix pyright errors
