+++ pokemon_ai/simulator/battle.py
@@ -1,12 +1,12 @@
 from __future__ import annotations
 
+import logging
 from random import random
 from typing import Optional
 
-from pokemon_ai.logger import log
 from pokemon_ai.simulator.damage import calculate_damage
 from pokemon_ai.simulator.moves import Move
-from pokemon_ai.simulator.player import Action, ActionChangeTo, ActionSelectMove, Player
+from pokemon_ai.simulator.player import Action, Player
 from pokemon_ai.simulator.pokedex import Pokemon
 
 
@@ -24,63 +24,60 @@ class Battle:
 
         if self.player1.get_active_pokemon().actual_hp <= 0:
             action = self.player1.choose_action_on_pokemon_dead(self.player2)
-            self.player1.change_pokemon_index_to(action.change_to)
-            log(f"{self.player1} changed to {action.change_to}")
+            self.player1.validate_change(action.value)
+            self.player1.active_pokemon_index = action.value
             return action, None
         if self.player2.get_active_pokemon().actual_hp <= 0:
             action = self.player2.choose_action_on_pokemon_dead(self.player1)
-            self.player2.change_pokemon_index_to(action.change_to)
-            log(f"{self.player2} changed to {action.change_to}")
+            self.player2.validate_change(action.value)
+            self.player2.active_pokemon_index = action.value
             return None, action
 
         action1 = self.player1.choose_action(self.player2)
         action2 = self.player2.choose_action(self.player1)
 
         # Handle pokemon change
-        if isinstance(action1, ActionChangeTo):
-            if action1.change_to not in self.player1.get_available_pokemons_for_change():
-                raise Exception(
-                    f"{self.player1} tried to change to {action1.change_to} but it's not available"
-                )
-            self.player1.change_pokemon_index_to(action1.change_to)
-            log(f"{self.player1} changed to {action1.change_to}")
-        if isinstance(action2, ActionChangeTo):
-            if action2.change_to not in self.player2.get_available_pokemons_for_change():
-                raise Exception(
-                    f"{self.player2} tried to change to {action2.change_to} but it's not available"
-                )
-            self.player2.change_pokemon_index_to(action2.change_to)
-            log(f"{self.player2} changed to {action2.change_to}")
+        if action1.is_change():
+            self.player1.validate_change(action1.value)
+            self.player1.active_pokemon_index = action1.value
+            logging.info(f"{self.player1} changed their pokemon")
+        if action2.is_change():
+            self.player2.validate_change(action2.value)
+            self.player2.active_pokemon_index = action2.value
+            logging.info(f"{self.player2} changed their pokemon")
 
         active_pokemon1 = self.player1.get_active_pokemon()
         active_pokemon2 = self.player2.get_active_pokemon()
 
         # Handle pokemon damage
-        if isinstance(action1, ActionSelectMove) and isinstance(action2, ActionSelectMove):
+        if action1.is_move() and action2.is_move():
             # to handle both player choose a move
             c1, c2 = get_spe_ordered_pokemon(
-                (active_pokemon1, action1.move), (active_pokemon2, action2.move)
+                (active_pokemon1, active_pokemon1.actual_moves[action1.value - 6]),
+                (active_pokemon2, active_pokemon2.actual_moves[action2.value - 6]),
             )
-            log(f"{c1[0]} used {action1.move}!")
+            logging.info(f"{c1[0]} used {c1[1]}!")
             damage = calculate_damage(c1[0], c2[0], c1[1])
             c2[0].actual_hp -= damage
-            log(f"{c2[0]} got {damage}")
+            logging.info(f"{c2[0]} got {damage}")
             if c2[0].actual_hp > 0:
-                log(f"{c2[0]} used {action2.move}!")
+                logging.info(f"{c2[0]} used {c2[1]}!")
                 damage = calculate_damage(c2[0], c1[0], c2[1])
                 c1[0].actual_hp -= damage
-                log(f"{c1[0]} got {damage}")
+                logging.info(f"{c1[0]} got {damage}")
         else:
-            if isinstance(action1, ActionSelectMove):
-                log(f"{active_pokemon1} used {action1.move}!")
-                damage = calculate_damage(active_pokemon1, active_pokemon2, action1.move)
+            if action1.is_move():
+                move = active_pokemon1.actual_moves[action1.value - 6]
+                logging.info(f"{active_pokemon1} used {move}!")
+                damage = calculate_damage(active_pokemon1, active_pokemon2, move)
                 active_pokemon2.actual_hp -= damage
-                log(f"{active_pokemon2} got {damage}")
-            if isinstance(action2, ActionSelectMove):
-                log(f"{active_pokemon2} used {action2.move}!")
-                damage = calculate_damage(active_pokemon2, active_pokemon1, action2.move)
+                logging.info(f"{active_pokemon2} got {damage}")
+            if action2.is_move():
+                move = active_pokemon2.actual_moves[action2.value - 6]
+                logging.info(f"{active_pokemon2} used {move}!")
+                damage = calculate_damage(active_pokemon2, active_pokemon1, move)
                 active_pokemon1.actual_hp -= damage
-                log(f"{active_pokemon1} got {damage}")
+                logging.info(f"{active_pokemon1} got {damage}")
 
         return action1, action2
 
@@ -91,7 +88,7 @@ class Battle:
             return self.player1
 
     def __repr__(self) -> str:
-        return f"Battle({self.player1}, {self.player2})"
+        return f"Battle:\n  {self.player1}\n  {self.player2})"
 
     def validate(self):
         for Player in (self.player1, self.player2):
@@ -99,18 +96,18 @@ class Battle:
                 raise ValueError("Pokemon must have at least one move")
 
     def run(self) -> Player:
-        log(self)
+        logging.info(self)
         self.validate()
         while True:
-            log("")
+            logging.info("")
             self.forward_step()
             winner = self.get_winner()
             if winner is not None:
-                log(f"{winner} won the battle!")
+                logging.info(f"{winner} won the battle!")
                 return winner
             if self.turn > 500:
                 raise Exception("Battle is too long")
-            log(self)
+            logging.info(self)
 
     def to_array(self):
         return [*self.player1.to_array(), *self.player2.to_array()]
